pyproject.toml
[tool.poetry]
name = "curvature-estimation"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "curvature_estimation"}]

[tool.poetry.dependencies]
python = "3.11.0"
pymc = "^5.9.1"
jax = "^0.4.20"
ipykernel = "^6.26.0"
pandas = "^2.1.2"
scipy = "^1.11.3"
numba = "^0.58.1"
arviz = "^0.16.1"
sympy = "^1.12"
geomstats = "^2.7.0"
torch = "^2.1.1"
nbformat = "^5.9.2"
nbclient = "^0.9.0"
jaxlib = "^0.4.22"
blackjax = "^1.0.0"
numpyro = "^0.13.2"
autograd = "^1.6.2"
torchdiffeq = "^0.2.3"
seaborn = "^0.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
